+++ src/Package/InstallFile/Writer/InstallFileJsonWriter.php
@@ -13,7 +13,7 @@ namespace Puli\RepositoryManager\Package\InstallFile\Writer;
 
 use Puli\RepositoryManager\IOException;
 use Puli\RepositoryManager\Package\InstallFile\InstallFile;
-use Puli\RepositoryManager\Package\InstallFile\PackageMetadata;
+use Puli\RepositoryManager\Package\PackageMetadata;
 use Symfony\Component\Filesystem\Filesystem;
 use Webmozart\Json\JsonEncoder;
 use Webmozart\PathUtil\Path;